backend/requirements.txt

Flask
flask-cors
mysql-connector-python
python-dotenv
yfinance
